#1156 [REFACTOR] OptionModal 네이밍 IconOptionModal로 변경
