+++ src/shared/component/Modal/IconOptionModal/IconOptionModal.module.css
