+++ src/shared/component/index.js
@@ -34,7 +34,7 @@ export { default as DimModalLayout } from './Modal/DimModalLayout/DimModalLayout
 export { default as ConfirmModal } from './Modal/ConfirmModal/ConfirmModal';
 export { default as NoticeModal } from './Modal/NoticeModal/NoticeModal';
 export { default as MoreOptionModal } from './Modal/MoreOptionModal/MoreOptionModal';
-export { default as OptionModal } from './Modal/OptionModal/OptionModal';
+export { default as IconOptionModal } from './Modal/IconOptionModal/IconOptionModal';
 
 // suspense
 export { default as ServerErrorFallback } from './suspense/ServerErrorFallback/ServerErrorFallback';
